requirements.txt
langchain==0.1.16
langchain-community==0.0.32
langchain-huggingface==0.0.3
streamlit==1.32.2
sentence-transformers==2.6.1
faiss-cpu==1.9.0.post1
PyPDF2==3.0.1
python-dotenv==1.0.1
huggingface_hub>=0.24.0
torch==2.6.0
torchvision==0.21.0
transformers==4.48.2
requests>=2.31.0
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0